Função para fazer parse do json dos prêmios
